+++ user-profile/src/main/java/ru/walkername/user_profile/events/RatingCreated.java
@@ -2,34 +2,34 @@ package ru.walkername.user_profile.events;
 
 public class RatingCreated {
 
-    private int userId;
+    private Long userId;
 
-    private int movieId;
+    private Long movieId;
 
     private int rating;
 
     public RatingCreated() {
     }
 
-    public RatingCreated(int userId, int movieId, int rating) {
+    public RatingCreated(Long userId, Long movieId, int rating) {
         this.userId = userId;
         this.movieId = movieId;
         this.rating = rating;
     }
 
-    public int getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(int userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 
-    public int getMovieId() {
+    public Long getMovieId() {
         return movieId;
     }
 
-    public void setMovieId(int movieId) {
+    public void setMovieId(Long movieId) {
         this.movieId = movieId;
     }
